Crate bracket_algorithm_traits

This crate is part of the bracket-lib family.

It provides traits for you to implement on your map (and other geometric constructs), translating them into a format that works with the bracket-pathfinding and bracket-geometry systems.

It is a separate crate so that both can depend upon it.

Defaults are provided to get you up and running quickly, but may (should!) be overridden if you don’t want to use my default array striding.

For example:

use bracket_algorithm_traits::prelude::{BaseMap, Algorithm2D};
use bracket_geometry::prelude::Point;

struct TestMap{};
impl BaseMap for TestMap {}
impl Algorithm2D for TestMap{
    fn dimensions(&self) -> Point {
        Point::new(2, 2)
    }
}

Modules

prelude

Exported traits